backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
redis==5.2.1
celery==5.4.0
boto3==1.36.2
jpype1==1.5.1
mpxj==12.9.5
python-dotenv==1.0.1
websockets==14.1
ypy-websocket==0.14.1
httpx==0.28.1
bcrypt==4.2.1
emails==0.6
jinja2==3.1.5
